Introspection is expensive and the number of persistent classes is small, so we'll cache the introspected primary key fields and reuse them. This has the added benefit that we can eliminate storing the key fields entirely because we know they must be exactly the set of primary key fields for the object which never changes.
We will have to ensure that we flush our caches whenever we do a new release,
but we would have to do that anyway for a variety of reasons. Actually, we
could perhaps make use of the persistent object's versions and note the version
number of the cached instances in the cache and ignore the cache if they
differ.