Well, OK. We're going to take the pain and break binary compatibility and use thsi opportunity to move a whole heap of stuff into impl that doesn't really need to be visible to external parties. We're changing the return value of all expression creating factory classes to SQLExpression or FluentExp because those expose all the functionality one might need on the returned values and there's otherwise no value to be had getting an And() or a LiteralExp() and a downside of preventing us from replacing the implementation of those expressions without breaking binary compatibility.
I hemmed a bit about everything in depot.operator. In theory everything there
is available via a combination of Exps and FluentExp but in practice Whirled
uses a ton of those classes directly (whre a ton is defined as 64 imports of
classes in depot.operator). Perhaps I'll go on a cleanup jihad there and then
move all that to impl as well.

The rabbit holes, they always go deep.