Accept Iterable rather than Collection everywhere we reasonably can.
Note one potentially consequential diff in DepotRepository:

-        return keys.isEmpty() ? Collections.<T>emptyList() :
-            _ctx.invoke(new FindAllQuery.WithKeys<T>(_ctx, keys));
+        return _ctx.invoke(new FindAllQuery.WithKeys<T>(_ctx, keys));

This should not actually impact anything. FindAllQuery.WithKeys first checks
for keys that are cached, coping with a potentially empty set (there will be
none), then if there are keys that remain to be loaded from the database, it
loads them. The first check will NOOP as will the second check. So this should
just be a slightly more circuitous path to getting back an empty collection. It
should not result in a zero row matching query being sent to the database.

With that in mind, it seemed inelegant to prematurely optmize the return of an
empty collection.