Properly ignore the last argument if there are an odd number.
When we saw the problem, I thought "there's no way!" because
the Logger is hardened and battle-tested.

But no. Since depot no longer depends on samskivert, it has
reimplemented Logger and throws away a lot of the goodness that
has evolved over the years. Goodness like:
- warning if there are an odd number of args and the last one
  isn't a Throwable.
- catching problems toString()ing any of the objects in the message.

Let's go ride a fixie.