Use DriverManager.getConnection().
This is the preferred way to obtain a JDBC connection in this millenium. No
longer does one manually instantiate the JDBC driver by name. The JDBC driver
jar file will have a mapping that indicates that it handles a particular type
of JDBC URL and the JDBC infrastructure will handle instantiating the driver.